Write -ten as -Z
These words have to be added to the dictionary manually though. I went
to the dictionary to add them for all words on -t admitting a -ten
ending.

This gives a few conflicts, such as kannen / kanten, in which case the
-en ending "wins". So, kanten needs to be written in two strokes.

Besides this, I added many -t forms that were still missing.

Closes #3.