Add support for Gemini protocol
So now this works with the Georgi.

Also removed -D and -G as suffix keys, because they don't make sense and
conflict with valid words. (Not sure why they were suffix keys in the
first place. Well, for -D it *would* make sense if it could output {^d},
but unfortunately that doesn't seem possible in Plover.)