Fix license in setup.cfg
The license mentioned in setup.cfg (GPL2+) differs from the LICENSE file (CC-BY 4.0). The idea was to have the code covered by the GPL, and the dictionary covered by CC-BY. However, this is just confusing in practice, and there is not that much code anyway, so I am putting everything under CC-BY now.

Fixes #1.